+++ src/lightning/dtos/lightning-info.dto.ts
@@ -1,7 +1,6 @@
-// import { PickType } from '@nestjs/swagger';
-// import { LightningInfoEntity } from '../entities/lightning.entity';
-
-// export class CreateLightningInfoDto extends PickType(LightningInfoEntity, [
-//   'writer_id',
-//   'meeting_date',
-// ] as const) {}
+class CreateLightningDto {
+  writer_id: number;
+  title: string;
+  contents: string;
+  meeting_date: Date;
+}
