+++ src/lightning/lightning.module.ts
@@ -2,7 +2,7 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { Module } from '@nestjs/common';
 import { LightningController } from './controllers/lightning.controller';
 import { LightningService } from './services/lightning.service';
-import { LightningInfoEntity } from './entities/lightning.entity';
+import { LightningInfoEntity } from './entities/lightning-info.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([LightningInfoEntity])],
