+++ src/lightning/services/lightning.service.ts
@@ -5,7 +5,7 @@ import { Injectable, UnauthorizedException } from '@nestjs/common';
 import {
   LightningBoardEntity,
   LightningInfoEntity,
-} from '../entities/lightning.entity';
+} from '../entities/lightning-info.entity';
 import { InjectRepository } from '@nestjs/typeorm';
 
 @Injectable()
@@ -15,15 +15,15 @@ export class LightningService {
     private readonly ligntningRepository: Repository<LightningInfoEntity>, // private readonly configService: ConfigService,
   ) {}
 
-  async createPost(active, meeting_date) {
-    // const user = await this.ligntningRepository.findOne({ meeting_date });
-    // if (user) {
-    //   throw new UnauthorizedException('해당하는 이메일은 이미 존재합니다.');
-    // }
+  async createPost(createLightningDto: CreateLightningDto) {
+    const { meeting_date } = createLightningDto;
+    const user = await this.ligntningRepository.findOne({ meeting_date });
+    if (user) {
+      throw new UnauthorizedException('해당하는 이메일은 이미 존재합니다.');
+    }
 
     await this.ligntningRepository.save({
-      active,
-      meeting_date,
+      ...createLightningDto,
     });
 
     return { success: true };
